+++ protocol/contracts/liquidity/GasOptimizedFlashLoanAggregatorProxy.sol
@@ -7,25 +7,23 @@ import "./IFlashLoanAggregator.sol";
 import "./QuickSort.sol";
 
 contract GasOptimizedFlashLoanAggregatorProxy is FlashLoanAggregatorProxy {
-
 	constructor(IFlashLoanAggregator _aggregator) FlashLoanAggregatorProxy(_aggregator, true) {}
 
 	function sortLenders(IERC3156FlashLender[] memory _lenders, address _token) internal override {
-		int[] memory liquidities = new int[](_lenders.length);
-		uint[] memory indexes = new uint[](_lenders.length);
+		int256[] memory liquidities = new int256[](_lenders.length);
+		uint256[] memory indexes = new uint256[](_lenders.length);
 
-		for (uint i = 0; i < _lenders.length; i++) {
+		for (uint256 i = 0; i < _lenders.length; i++) {
 			indexes[i] = i;
-			liquidities[i] = -int(_lenders[i].maxFlashLoan(_token));
+			liquidities[i] = -int256(_lenders[i].maxFlashLoan(_token));
 		}
 
 		QuickSort.quickSortByValues(indexes, liquidities);
 
 		delete lenders[_token];
 
-		for (uint i = 0; i < _lenders.length; i++) {
+		for (uint256 i = 0; i < _lenders.length; i++) {
 			lenders[_token].push(_lenders[indexes[i]]);
 		}
 	}
-
 }
